Criando Seeders

+++ database/migrations/2023_02_20_122317_create_alocacoes_table.php
@@ -14,8 +14,8 @@ class CreateAlocacoesTable extends Migration
     public function up()
     {
         Schema::create('alocacoes', function (Blueprint $table) {
-            $table->integer('desenvolvedor_id')->unsigned();
-            $table->integer('projeto_id')->unsigned();
+            $table->bigInteger('desenvolvedor_id')->unsigned();
+            $table->bigInteger('projeto_id')->unsigned();
             $table -> primary(['projeto_id','desenvolvedor_id']);
             $table -> foreign('desenvolvedor_id')->references('id')->on('desenvolvedores');
             $table-> foreign("projeto_id")->references('id')-> on('projetos');
